seed修正
画像がない場合のエラー修正え

+++ db/seeds.rb
@@ -51,9 +51,12 @@ CSV.foreach('db/csv/shop.csv', headers: true) do |row|
     genre_id: row['genre_id'],
     area_id: row['area_id']
   )
-  shop.images.attach(io: File.open("db/images/shops/shop#{n}_1.jpg"), filename: "shop#{n}_1.jpg", content_type: 'application/jpg')
-  shop.images.attach(io: File.open("db/images/shops/shop#{n}_2.jpg"), filename: "shop#{n}_2.jpg", content_type: 'application/jpg')
-  shop.images.attach(io: File.open("db/images/shops/shop#{n}_3.jpg"), filename: "shop#{n}_3.jpg", content_type: 'application/jpg')
+  shop_image1 = "db/images/shops/shop#{n}_1.jpg"
+  shop_image2 = "db/images/shops/shop#{n}_2.jpg"
+  shop_image3 = "db/images/shops/shop#{n}_3.jpg"
+  shop.images.attach(io: File.open("db/images/shops/shop#{n}_1.jpg"), filename: "shop#{n}_1.jpg", content_type: 'application/jpg') if File.exist?(shop_image1)
+  shop.images.attach(io: File.open("db/images/shops/shop#{n}_2.jpg"), filename: "shop#{n}_2.jpg", content_type: 'application/jpg') if File.exist?(shop_image2)
+  shop.images.attach(io: File.open("db/images/shops/shop#{n}_3.jpg"), filename: "shop#{n}_3.jpg", content_type: 'application/jpg') if File.exist?(shop_image3)
   shop.save!
   n += 1
 end
